requirements.txt
streamlit>=1.24.0
fastapi>=0.95.0
requests>=2.28.0
uvicorn>=0.21.0
redis>=4.5.0
starlette>=0.26.1
python-dotenv>=1.0.0
httpx>=0.24.0
kafka-python>=2.0.0